Simple chat filter bot using logistic regression
Training, crossvalidation and test data can be found from data.py.
Data features are list of most commonly found words in training data.
Data output is boolean value if sentence is filtered or not.

Simple logistic regression is used to calculate optimal theta values which
are then used to calculate the probability if new sentence is filtered or
not. If the probability is greater than 0.5 then it results as filtered
sentence.